Moved injection token to news.service.ts
