+++ src/app/app.module.ts
@@ -3,13 +3,11 @@ import { NgModule } from '@angular/core';
 import { HttpClientModule } from '@angular/common/http';
 import { FormsModule } from '@angular/forms';
 
-import { NEWS_SERVICE } from './injection-tokens';
-
 import { AppComponent } from './app.component';
 import { NewsComponent } from './components/news/news.component';
 import { RegionSelectComponent } from './components/region-select/region-select.component';
 
-import { SmartableAiNewsService } from './services/news.service';
+import { NEWS_SERVICE, SmartableAiNewsService } from './services/news.service';
 
 @NgModule({
   declarations: [
