--- src/app/injection-tokens.ts
@@ -1,5 +1,0 @@
-import { InjectionToken } from '@angular/core';
-
-import { NewsService } from './services/news.service';
-
-export const NEWS_SERVICE = new InjectionToken<NewsService>('NewsService');
